msfdrb: add module_create method.
When we run @framework.modules.create(module_name) in the drb client
manually and not on the drb server the garbage collector on the drb
server removes the object and we can not call any methods on it any
more. To prevent the gc from removing the object we hold it in a hash
on the server, so there is a reference on it and the gc does not remove
it. When not doing it like this the autopwn fails in 30% the cases.