autopwn: The autopwn is working and integrated now.
When calling action.attack_host the host will be attacked now. Now it
is possible to attack a host like an old windows xp which is vulnerable
to the ms08_067_netapi exploit. The Code is based on msf-autopwn which
is based on the autopwn component of metasploit.

There are some TODOS missing:
 * make it multi threaded
 * work correctly with pivoted host (untested)
 * run in background, now action.attack_host blocks
 * update and extend documentation
 * add some tests