drb: send a proxy over the drb connection and not a serialized object.
When a serialized object is send over the drb connection the class it
was created from is needed on the client side. As we do not want to
require the complete msf framework, proxy methods should be used. To do
so we should include DRbUndumped in all classes, but that will be not
good to do in the complete msf framework. Now we changed the drb dump
method where the check for DRbUndumped is done to also send proxies if
some msf classes are used.