tests: add mock for msf and some part of the fidius core.
The tests are now divided into unit tests in test/fidius/ and
integration test in test/integration/. The unit tests do not need a
running msf framework or anything else they are just using the mocked
objects. The integration tests need a special setup e.g. a vm running
with windows to attack and so on.