postexploit: add an optional args parameter to all post exploits
The info of the post exploits now contains a attribute args which
contains the possible extra options for this post exploits. The user
should choose one of these, if it is available. This makes it possible
to specify the host, the browser autopowner should forward to when
bubbtrapping the web server.